[TASK] Controller refactoring
Refs #495
